+++ src/Business/BookMark/BookMarkAllRemove.ts
@@ -1,1 +1,3 @@
+type propsType = {};
+
 export const removeAllItem = () => {};
