+++ src/fetch/common/axiosApi.ts
@@ -1,14 +1,19 @@
 import axios from "axios";
+import { Cookies } from "react-cookie";
+
+const cookies = new Cookies();
+
 axios.defaults.withCredentials = true;
 const baseAPI = (url = "https://www.officefinder.site/", options?: any) => {
   return axios.create({ baseURL: url, ...options });
 };
 const authAPI = (url = "https://www.officefinder.site/", options?: any) => {
-  const token = localStorage.getItem("jwt_token");
+  const token = cookies.get("Authorization");
   return axios.create({
     baseURL: url,
     headers: {
-      Authorization: `bearer ${token}`,
+      "Accept": "application/json",
+      Authorization: `${token}`,
     },
     ...options,
   });
