+++ src/fetch/get/agent.ts
@@ -14,22 +14,5 @@ export const fetchReviewsData = async (): Promise<MyOfficeResponse> => {
   return response.data;
 };
 
-// 지금은 모두 같지만 추후 url을 변경해야합니다.
-
-export const fetchBookMarkData = () => authInstance.get("api/bookmarks").then(res => res.data);
-
-export const fetchMyBookingsData = () =>
-  axios.get("https://my-json-server.typicode.com/fefdfea1/jsonData/db").then(res => res.data);
-
-export const fetchBookingData = () =>
-  baseInstance
-    .get("api/customers/info/leases", {
-      headers: {
-        "Access-Control-Allow-Origin": `http://127.0.0.1:5173`,
-        "Access-Control-Allow-Credentials": "true",
-      },
-    })
-    .then(res => res.data);
-
-export const fetchMyPageData = () =>
-  axios.get("https://my-json-server.typicode.com/fefdfea1/jsonData2/db").then(res => res.data);
+// export const fetchMyPageData = () => authInstance.get("api/customers/info").then(res => res.data);
+export const fetchMyPageData = () => authInstance.get("api/customers/info").then(res => res.data);
