+++ src/fetch/post/agent.ts
@@ -12,31 +12,3 @@ export const postNewOfficeData = async (newOfficeData: NewOfficePost) => {
     throw error;
   }
 };
-
-export const deleteBookmark = (officeId: number, cookie: string) => {
-  axios.post(
-    "/api/bookarks/delete",
-    {
-      officdId: officeId,
-    },
-    {
-      headers: {
-        Authorization: cookie,
-      },
-    },
-  );
-};
-
-export const addBookMark = (officeId: number, cookie: string) => {
-  axios.post(
-    "/api/bookmarks/submit",
-    {
-      officdId: officeId,
-    },
-    {
-      headers: {
-        Authorization: cookie,
-      },
-    },
-  );
-};
